requirements.txt
streamlit
google-generativeai
pypdf
python-docx